+++ doorstop/core/publisher.py
@@ -6,7 +6,8 @@ import logging
 
 import markdown
 
-from doorstop.common import DoorstopError
+from doorstop.common import DoorstopError, create_dirname
+from doorstop.core.types import iter_items
 from doorstop import settings
 
 CSS = os.path.join(os.path.dirname(__file__), 'files', 'doorstop.css')
@@ -23,22 +24,16 @@ def publish(obj, path, ext=None, **kwargs):
     @raise DoorstopError: for unknown file formats
 
     """
+    # Determine the output format
     ext = ext or os.path.splitext(path)[-1]
-    if ext in FORMAT:
+    check(ext)
 
-        # Create output directory
-        dirpath = os.path.dirname(path)
-        if not os.path.isdir(dirpath):
-            logging.info("creating {}...".format(dirpath))
-            os.makedirs(dirpath)
-
-        # Publish report
-        logging.info("publishing {}...".format(path))
-        with open(path, 'w') as outfile:  # pragma: no cover (integration test)
-            for line in lines(obj, ext, **kwargs):
-                outfile.write(line + '\n')
-    else:
-        raise DoorstopError("unknown format: {}".format(ext))
+    # Publish content to the specified path
+    create_dirname(path)
+    logging.info("creating file {}...".format(path))
+    with open(path, 'w') as outfile:  # pragma: no cover (integration test)
+        for line in lines(obj, ext, **kwargs):
+            outfile.write(line + '\n')
 
 
 def index(directory, extensions=('.html',)):
@@ -93,11 +88,9 @@ def lines(obj, ext='.txt', **kwargs):
     @raise DoorstopError: for unknown file formats
 
     """
-    if ext in FORMAT:
-        logging.debug("yielding {} as lines of {}...".format(obj, ext))
-        yield from FORMAT[ext](obj, **kwargs)
-    else:
-        raise DoorstopError("unknown format: {}".format(ext))
+    gen = check(ext)
+    logging.debug("yielding {} as lines of {}...".format(obj, ext))
+    yield from gen(obj, **kwargs)
 
 
 def lines_text(obj, indent=8, width=79):
@@ -110,7 +103,7 @@ def lines_text(obj, indent=8, width=79):
     @return: iterator of lines of text
 
     """
-    for item in _get_items(obj):
+    for item in iter_items(obj):
 
         level = _format_level(item.level)
 
@@ -173,7 +166,7 @@ def lines_markdown(obj):
     @return: iterator of lines of text
 
     """
-    for item in _get_items(obj):
+    for item in iter_items(obj):
 
         heading = '#' * item.depth
         level = _format_level(item.level)
@@ -215,19 +208,6 @@ def lines_markdown(obj):
                     yield '*' + slinks + '*'
 
         yield ""  # break between items
-
-
-def _get_items(obj):
-    """Get an iterator of items from from an item, list, or document."""
-    if hasattr(obj, 'items'):
-        # a document
-        return (i for i in obj.items if i.active)
-    try:
-        # an iterable
-        return iter(obj)
-    except TypeError:
-        # an item
-        return [obj]  # an item
 
 
 def _format_level(level):
@@ -283,6 +263,23 @@ def lines_html(obj):
 
 
 # Mapping from file extension to lines generator
-FORMAT = {'.txt': lines_text,
-          '.md': lines_markdown,
-          '.html': lines_html}
+FORMAT_LINES = {'.txt': lines_text,
+                '.md': lines_markdown,
+                '.html': lines_html}
+
+
+def check(ext):
+    """Confirm an extension is supported for publish.
+
+    @raise DoorstopError: for unknown formats
+
+    @return: lines generator if available
+
+    """
+    try:
+        return FORMAT_LINES[ext]
+    except KeyError:
+        exts = ', '.join(ext for ext in FORMAT_LINES)
+        msg = "unknown publish format: {} (options: {})".format(ext, exts)
+        exc = DoorstopError(msg)
+        raise exc from None